# Some easy micro optimizations to collection