Collectively, on this benchmark run with `--collect-only`:

```py
import pytest
@pytest.mark.parametrize("x", range(50000))
def test_foo(x): pass
```

Before:

```
============================ no tests ran in 11.37s ============================
         23775445 function calls (22920574 primitive calls) in 11.648 seconds
```

After:

```
============================ no tests ran in 10.51s ============================
         21725509 function calls (20870638 primitive calls) in 10.931 seconds
```

The first commit utilizes the fact that dicts are insertion-order preserving on Python>=3.7 to avoid  OrderedDict overhead when possible.

The second and third commits apply some micro optimizations to Node.

The fourth applies some minor optimizations to `PythonCollector.collect`.