@bluetech 2020-05-06
I scaled back the last commit a bit after the CI kindly informed me that I had missed some details :smile: Doesn't affect the benchmark results however.