@boxed 2020-05-12
I had some time to check this and the other changes that have landed in master. I am comparing 5.4.1 and master:

![before](https://user-images.githubusercontent.com/332428/81665266-cc9ed300-9440-11ea-86cc-00775a7090af.png)
![after](https://user-images.githubusercontent.com/332428/81665270-cdd00000-9440-11ea-9f2c-ac65eaebbdda.png)

It's slightly better but it's still very bad.