@boxed 2020-05-12
I tried pytest 3.10.1 and it's at 7.8s for the parametrize_huge benchmark. Again still very bad, but a lot better than pytest 5. I even checked 3.6 because I have a memory of this being faster on our suite at work: 7.5s so basically the same.